fix: update size type to int64_t and initialize struct stat in UABPackager and Builder
Changed the size type from int to int64_t in both Builder and UABPackager to ensure proper handling of larger values. Additionally, initialized struct stat variables to improve code clarity and maintainability.

+++ libs/linglong/src/linglong/builder/linglong_builder.cpp
@@ -1371,7 +1371,7 @@ utils::error::Result<void> Builder::commitToLocalRepo() noexcept
         if (!ret) {
             return LINGLONG_ERR(ret);
         }
-        info.size = static_cast<int>(*ret);
+        info.size = static_cast<int64_t>(*ret);
 
         QFile infoFile{ moduleOutput.filePath("info.json") };
         if (!infoFile.open(QIODevice::WriteOnly | QIODevice::Truncate)) {
